+++ test-sql/tsql/00/test004.sql
@@ -1,0 +1,9 @@
+CREATE TABLE Production.Categories
+(
+  categoryid   INT           NOT NULL IDENTITY,
+  categoryname NVARCHAR(15)  NOT NULL,
+  description  NVARCHAR(200) NOT NULL,
+  CONSTRAINT PK_Categories PRIMARY KEY(categoryid)
+);
+
+CREATE INDEX categoryname ON Production.Categories(categoryname);