+++ test-sql/tsql/00/test008.sql
@@ -1,0 +1,32 @@
+CREATE TABLE Sales.Orders
+(
+  orderid        INT          NOT NULL IDENTITY,
+  custid         INT          NULL,
+  empid          INT          NOT NULL,
+  orderdate      DATETIME     NOT NULL,
+  requireddate   DATETIME     NOT NULL,
+  shippeddate    DATETIME     NULL,
+  shipperid      INT          NOT NULL,
+  freight        MONEY        NOT NULL
+    CONSTRAINT DFT_Orders_freight DEFAULT(0),
+  shipname       NVARCHAR(40) NOT NULL,
+  shipaddress    NVARCHAR(60) NOT NULL,
+  shipcity       NVARCHAR(15) NOT NULL,
+  shipregion     NVARCHAR(15) NULL,
+  shippostalcode NVARCHAR(10) NULL,
+  shipcountry    NVARCHAR(15) NOT NULL,
+  CONSTRAINT PK_Orders PRIMARY KEY(orderid),
+  CONSTRAINT FK_Orders_Customers FOREIGN KEY(custid)
+    REFERENCES Sales.Customers(custid),
+  CONSTRAINT FK_Orders_Employees FOREIGN KEY(empid)
+    REFERENCES HR.Employees(empid),
+  CONSTRAINT FK_Orders_Shippers FOREIGN KEY(shipperid)
+    REFERENCES Sales.Shippers(shipperid)
+);
+
+CREATE NONCLUSTERED INDEX idx_nc_custid         ON Sales.Orders(custid);
+CREATE NONCLUSTERED INDEX idx_nc_empid          ON Sales.Orders(empid);
+CREATE NONCLUSTERED INDEX idx_nc_shipperid      ON Sales.Orders(shipperid);
+CREATE NONCLUSTERED INDEX idx_nc_orderdate      ON Sales.Orders(orderdate);
+CREATE NONCLUSTERED INDEX idx_nc_shippeddate    ON Sales.Orders(shippeddate);
+CREATE NONCLUSTERED INDEX idx_nc_shippostalcode ON Sales.Orders(shippostalcode);