+++ test-sql/tsql/00/test013.sql
@@ -1,0 +1,13 @@
+CREATE VIEW Sales.OrderTotalsByYear
+  WITH SCHEMABINDING
+AS
+
+SELECT
+  YEAR(O.orderdate) AS orderyear,
+  SUM(OD.qty) AS qty
+FROM Sales.Orders AS O
+  JOIN Sales.OrderDetails AS OD
+    ON OD.orderid = O.orderid
+GROUP BY YEAR(orderdate);
+GO
+