+++ test-sql/tsql/04/test001.sql
@@ -1,0 +1,13 @@
+SET NOCOUNT ON;
+USE master;
+IF DB_ID('Performance') IS NULL
+  CREATE DATABASE Performance;
+GO
+USE Performance;
+GO
+
+-- Creating and Populating the Nums Auxiliary Table
+SET NOCOUNT ON;
+IF OBJECT_ID('dbo.Nums', 'U') IS NOT NULL
+  DROP TABLE dbo.Nums;
+CREATE TABLE dbo.Nums(n INT NOT NULL PRIMARY KEY);