+++ test-sql/tsql/04/test010.sql
@@ -1,0 +1,8 @@
+SELECT
+  object_name,
+  counter_name,
+  instance_name,
+  cntr_value,
+  cntr_type
+FROM sys.dm_os_performance_counters;
+