+++ RELEASES.md
@@ -14,48 +14,83 @@
 
 ## Next Releases:
 
+### Version "Ararat" 0.0.37 (20.01.2015)
+
+* DECLARE @a = expr, @b = expr
+* SET a->b = 10;
+* VIEWS
+* WITH
+* Float numbers like 10e20
+* Wikibooks SQL samples as tests
+* Google Spreadsheet Tabletop (https://github.com/jsoma/tabletop) FROM-function
+ * alasql("SELECT * FROM GSHEET(?,{})",[key],function(){});
+
 ### Version "Armenia" 0.0.40 (20.01.2015)
 
 * T-SQL compatibility and multiple SQL Tests
+* Autoincrement
+* If-then-else JISON problem
+* UNION ORDER BY bug
 * Documentation
 * API documentation with JsDoc
 * CSV bug and headers
-* If-then-else JISON problem
-* UNION ORDER BY bug
+* SELECT INDEX, COLUMN, VALUE on column
 * WorkerJS for Node (primarily for test)
 
 ### Version "Sigirya" 0.0.45 (01.02.2015)
 
-* packege.json
-* Constraint and Foreign Keys
+* package.json
+* Datatypes + CONVERT
+* Functions
+* Constraint and Foreign Keys, indexes, Check, Default
 * CONTINUE/BREAK/FOR/LOOP
 
 ### Version "Elbrus" 0.0.50 (10.02.2015)
 
 * Transactions
+* SET NOCOUNT 
 
 ### Version "Prague" 0.0.55 (20.02.2015)
 
 * PIVOT/UNPIVOT
 * CROSS APPLY/OUTER APPLY
-* VIEWS
 
 ### Version "Manila" 0.0.60 (01.03.2015)
  
 * GROUP BY HIERARCHY, TOTAL, DETAIL
+* OVER PARTITION
 
 ### Version "Menlo Park" 0.0.65 (10.03.2015)
 
+* CREATE FUNCTION, CREATE PROCEDURE, EXEC
 * Bugs
 * Documentation
 
 ### Version "Vorkuta" 0.0.70 (20.03.2015)
 
+* Fluent interface improvement
 * Bugs
 * Documentation
 
-### Version "First Pancake" 0.1.0 (01.04.2015)
+### Version "San Diego" 0.0.75 (01.04.2015)
 
+* NeDB, MongoDB, Redis, Cassandra NoSQL
+* Bugs
+* Documentation
+
+### Version "Voronezh" 0.0.80 (10.04.2015)
+
+* Compatibility
+
+### Version "Rio de Janeiro" 0.0.85 (20.04.2015)
+* Speed
+
+### Version "Quito" 0.0.90 (01.05.2015)
+* Memory 
+
+### Version "First Pancake" 0.1.0 (10.05.2015)
+
+* Bugs
 * Speed, memory, and size optimization 
 * Compatibility: browsers, mobiles, SQL databases